+++ action_server/src/robocorp/action_server/__init__.py
@@ -1,6 +1,6 @@
 from typing import List
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 version_info = [int(x) for x in __version__.split(".")]
 
 __all__: List[str] = []