+++ action_server/tests/action_server_tests/test_server_secrets/pack_encryption/hello_action.py
@@ -1,0 +1,12 @@
+from robocorp.actions import Secret, action
+
+
+@action
+def get_private(private_info: Secret) -> str:
+    """
+    Returns the value passed to the private key.
+
+    Returns:
+        The value of the private key.
+    """
+    return private_info.value