+++ actions/src/robocorp/actions/_lint_action.py
@@ -264,7 +264,7 @@ def _check_docstring_contents(
     if arguments.args:
         for arg in arguments.args:
             desc = param_name_to_description.pop(arg.arg, None)
-            if pm is not None and _is_managed_param(pm, arg.arg):
+            if pm is not None and _is_managed_param(pm, arg.arg, node=node):
                 continue
 
             if not desc:
@@ -285,7 +285,7 @@ def _check_docstring_contents(
 
 
 def iter_lint_errors(
-    action_contents_file: str | bytes, pm: Optional[PluginManager] = None
+    action_contents_file: Union[str, bytes], pm: Optional[PluginManager] = None
 ) -> Iterator[Error]:
     ast = ast_module.parse(action_contents_file, "<string>")
     for _stack, node in _iter_nodes(ast, recursive=False):