+++ meta/src/robocorp/_meta/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "2.0.1"
+__version__ = "2.0.2"
 version_info = [int(x) for x in __version__.split(".")]