+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.3.0 - 2024-04-10
+
+- `action-server package metadata` now includes information on the secrets required
+  for each action.
+- Passing secrets now works (requires `robocorp-actions=0.2.0`)
+
 ## 0.2.1 - 2024-04-04
 
 - `action-server package build` no longer includes the `.zip` being created in the